V4: new struct implemented
