+++ .gitignore
@@ -2,4 +2,3 @@
 .utpm
 .vscode
 utpm
-/test
