+++ src/commands/install.rs
@@ -7,8 +7,8 @@ use crate::{
             check_path_dir, check_path_file, d_packages, datalocalutpm, get_current_dir,
             get_ssh_dir,
         },
+        specs::{Extra, TypstConfig},
         state::{Error, ErrorKind, Responses, Result},
-        TypstConfig,
     },
 };
 use git2::{build::RepoBuilder, Cred, FetchOptions, RemoteCallbacks, Repository};
@@ -76,11 +76,7 @@ pub fn init(cmd: &InstallArgs, res: &mut Responses, i: usize) -> Result<bool> {
     let utpm = file.utpm;
     let namespace = utpm
         .clone()
-        .unwrap_or(crate::utils::Extra {
-            version: None,
-            namespace: Some("local".to_string()),
-            dependencies: None,
-        })
+        .unwrap_or(Extra::new(None, Some("local".to_string()), None))
         .namespace
         .unwrap_or("local".into());
