+++ tests/old/lib.typ
@@ -1,0 +1,1 @@
+// This file is generated by UTPM (https://github.com/ThumusLive/utpm)
